Move the default cur parameters into common
