+++ src/common/helpers/constants/DEFAULT_CURRENCY_PARAMETERS.ts
