+++ src/common/helpers/constants/index.ts
@@ -1,5 +1,7 @@
 import TRANSACTION from './TRANSACTION'
+import DEFAULT_CURRENCY_PARAMETERS from './DEFAULT_CURRENCY_PARAMETERS'
 
 export default {
   TRANSACTION,
+  DEFAULT_CURRENCY_PARAMETERS,
 }
