+++ src/front/shared/helpers/btc.ts
@@ -2,7 +2,7 @@ import * as bitcoin from 'bitcoinjs-lib'
 import { getState } from 'redux/core'
 import actions from 'redux/actions'
 import config from './externalConfig'
-import DEFAULT_CURRENCY_PARAMETERS from './constants/DEFAULT_CURRENCY_PARAMETERS'
+import DEFAULT_CURRENCY_PARAMETERS from 'common/helpers/constants/DEFAULT_CURRENCY_PARAMETERS'
 import constants from 'common/helpers/constants'
 import api from './api'
 import BigNumber from 'bignumber.js'
