+++ src/front/shared/helpers/constants/index.ts
@@ -5,7 +5,6 @@ import modals from './modals'
 import tradeTicker from './TRADE_TICKERS'
 import tokenDecimals from './TOKEN_DECIMALS'
 import pairTypes from './PAIR_TYPES'
-import defaultCurrencyParameters from './DEFAULT_CURRENCY_PARAMETERS'
 import minAmount from './minAmount'
 import minAmountOffer from './minAmountOffer'
 import coinsWithDynamicFee from './coinsWithDynamicFee'
@@ -19,7 +18,6 @@ export default {
   tokenDecimals,
   notifications,
   pairTypes,
-  defaultCurrencyParameters,
   modals,
   minAmount,
   minAmountOffer,
