+++ src/front/shared/helpers/ethToken.ts
@@ -1,6 +1,6 @@
 import config from './externalConfig'
 import eth from './eth'
-import DEFAULT_CURRENCY_PARAMETERS from './constants/DEFAULT_CURRENCY_PARAMETERS'
+import DEFAULT_CURRENCY_PARAMETERS from 'common/helpers/constants/DEFAULT_CURRENCY_PARAMETERS'
 import BigNumber from 'bignumber.js'
 
 const isEthToken = ({ name }) => Object.keys(config.erc20).includes(name.toLowerCase())
